+++ src/routes/Routes.kt
@@ -1,0 +1,18 @@
+package routes
+
+import io.ktor.http.content.resource
+import io.ktor.http.content.static
+import io.ktor.routing.Routing
+
+fun Routing.routes() {
+    // システム全般、サンプルコード
+    systems()
+
+    // メモ機能
+    memos()
+
+    // 静的アクセス
+    static("/") {
+        resource("favicon.ico")
+    }
+}
